Added class for equations
